theme: rename compat -> unified
